Update emotion detection, IdempotencyKey and code cleanups
Added UUID-based IdempotencyKey in HttpHeaders for authService and implemented retry functionality in noteService. Also introduced an AI-based emotion detection process for enhanced user experience. Code and components were cleaned and organized for better maintainability and performance.
