+++ app/controllers/NoteController.scala
@@ -6,19 +6,17 @@ import net.logstash.logback.argument.StructuredArguments._
 import org.slf4j.{Logger, LoggerFactory}
 import play.api.libs.json.{JsError, JsValue, Json}
 import play.api.mvc.{AbstractController, Action, AnyContent, ControllerComponents}
-import service.ai.{EmoDetectionServiceWithAssistant, EmotionDetectionService}
+import service.ai.EmotionDetectionService
 import service.model.DetectEmotionRequest
 import service.{NoteService, NoteTodoService}
 
 import javax.inject.Inject
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
-import scala.util.Try
 
 class NoteController @Inject()(cc: ControllerComponents,
                                noteService: NoteService,
                                emotionDetectionService: EmotionDetectionService,
-                               emotionDetectionServiceV2: EmoDetectionServiceWithAssistant,
                                noteTodoService: NoteTodoService,
                                authenticatedAction: AuthenticatedAction)
   extends AbstractController(cc){
@@ -39,21 +37,18 @@ class NoteController @Inject()(cc: ControllerComponents,
   }
 
   def detectEmotion(): Action[JsValue] = Action(parse.json) andThen authenticatedAction async { implicit token =>
-    logger.info("V1 Detecting emotion")
     token.body.validate[Note].fold(
       errors => {
         Future.successful(BadRequest(Json.obj("message" -> JsError.toJson(errors))))
       },
       note => {
+        token.headers.get("X-IdempotencyKey")
         val v1EmotionFuture = emotionDetectionService.detectEmotion(DetectEmotionRequest(note.text, token.user.userId))
-        val v2EmotionFuture = emotionDetectionServiceV2.detectEmotion(DetectEmotionRequest(note.text, token.user.userId))
-
-        Future.firstCompletedOf(Seq(v1EmotionFuture, v2EmotionFuture)).map { resp =>
+        v1EmotionFuture.map { resp =>
           Ok(Json.toJson(resp))
         }.recover {
           case e: Exception =>
-            logger.error(s"V1 Failed to detect emotion: $e", e)
-            InternalServerError(Json.obj("message" -> "V1 Failed to detect emotion"))
+            InternalServerError(Json.obj("message" -> "Failed to detect emotion"))
         }
       }
     )
