+++ app/service/ai/EmoDetectionServiceWithAssistantImpl.scala
@@ -1,24 +1,19 @@
 package service.ai
 
-import com.google.inject.ImplementedBy
 import dao.model.EmotionDetectionResult
 import play.api.libs.json.Json
 import service.model.DetectEmotionRequest
 
-import javax.inject.Inject
+import javax.inject.{Inject, Named}
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 import scala.util.Success
 
-@ImplementedBy(classOf[EmoDetectionServiceWithAssistantImpl])
-trait EmoDetectionServiceWithAssistant {
-  def detectEmotion(request: DetectEmotionRequest): Future[EmotionDetectionResult]
-}
-
+@Named("ChatGptAssistant")
 class EmoDetectionServiceWithAssistantImpl @Inject()(
                                                       aiAssistantService: AiAssistantService,
                                                       aiDbService: AiDbService
-                                                    ) extends EmoDetectionServiceWithAssistant {
+                                                    ) extends EmotionDetectionService {
   private val assistantType = "EmoDetection"
 
   private lazy val logger = play.api.Logger(getClass)
@@ -39,15 +34,15 @@ class EmoDetectionServiceWithAssistantImpl @Inject()(
     }
     responseFuture.onComplete {
       case scala.util.Success(_) =>
-        logger.info(s"Successfully detected emotion for request using V2, userId: ${request.userId}")
+        logger.info(s"V2 Successfully detected emotion for request, userId: ${request.userId}")
       case scala.util.Failure(e) =>
-        logger.error(s"Failed to detect emotion for request using V2, userId: ${request.userId}", e)
+        logger.error(s"V2 Failed to detect emotion for request, userId: ${request.userId}", e)
     }
     responseFuture.andThen {
       case Success(x) =>
         val endTime = System.nanoTime()
         val elapsedTime = (endTime - startTime) / 1e9d
-        logger.info(s"Total elapsed time for detectEmotion V2: $elapsedTime seconds")
+        logger.info(s"V2 Total elapsed time for detectEmotion: $elapsedTime seconds")
         aiDbService.saveAiResponse(request.userId, EmotionDetectionResult.emotionDetectionResultFormat.writes(x),
           Option(request.text), Option("emo detection v2"), Option(elapsedTime))
     }
