+++ ui/package.json
@@ -35,6 +35,7 @@
     "moment": "^2.29.4",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
+    "uuid": "^9.0.1",
     "zone.js": "~0.13.0"
   },
   "devDependencies": {
@@ -46,6 +47,7 @@
     "@types/d3-shape": "^3.1.1",
     "@types/jasmine": "~4.3.0",
     "@types/jsonwebtoken": "^9.0.2",
+    "@types/uuid": "^9.0.7",
     "dotenv": "^16.3.1",
     "jasmine-core": "~4.5.0",
     "jsonwebtoken": "^9.0.0",
